Move generic checklist code into its own package
