+++ checklist/checklist_item.go
@@ -1,4 +1,8 @@
-package wtf
+package checklist
+
+import (
+	"github.com/senorprogrammer/wtf/wtf"
+)
 
 // ChecklistItem is a module for creating generic checklist implementations
 // See 'Todo' for an implementation example
@@ -10,7 +14,7 @@ type ChecklistItem struct {
 // CheckMark returns the string used to indicate a ChecklistItem is checked or unchecked
 func (item *ChecklistItem) CheckMark() string {
 	if item.Checked {
-		return Config.UString("wtf.mods.todo.checkedIcon", "x")
+		return wtf.Config.UString("wtf.mods.todo.checkedIcon", "x")
 	} else {
 		return " "
 	}
